moved main activity db
